[finishes #162653251]Moved app errorhandlers to the run.py file
