+++ instance/config.py
@@ -1,13 +1,15 @@
 import os
-
+import datetime
 class Config(object):
     """Parent configuration class."""
     DEBUG = False
     CSRF_ENABLED = True
-    SECRET = os.getenv('SECRET')
+    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY')
+    time = os.getenv('JWT_ACCESS_TOKEN_EXPIRES')
     DB_HOST = os.getenv('DB_HOST')
     DB_USER = os.getenv('DB_USER')
     DB_PASSWORD = os.getenv('DB_PASSWORD')
+    JWT_ACCESS_TOKEN_EXPIRES = datetime.timedelta(hours=int(time))
    
 
 class DevelopmentConfig(Config):
